Add help and footer region. Add full page support.
Adjust CSS styles to allow main body to fill the entire page when the left
sidebar region is hidden.

Needed to override the default panels-pane template in order to set
the pane title to use an H4 tag rather than an H2. Is there really no
other way to do this?

Closes #22