Initial commit of changes for implementing WSDOT home page.
Closes #44

Replaced High Impact Alerts with a custom block.

If I am using a template to override the look and feel of the panel pane
it needs to be an existing custom block otherwise the id is not fixed and
the custom template will be applied to another pane when the High Impact
Alerts block is hidden.