Style modifications for calendar view size when empty.
When there were no entries in the calender month view the table-cell containing
the calendar shrunk to 230px. Inheriting the width on the #main id appears to
fix it.

Also added a style to override padding from the item-list class which was
causing the controls to push off the right of the page.

Closes #47