Use Panelizer to layout and style the project content type.
There is still some style changes to make, however, the layout
no longer depends on overridden templates in our custom there base.
Using Panelizer means we can add and remove content boxes and play
with the style using the UI.

Removed dependencies on Display Suite since I am no longer using it.
This shouldn't break anything, however, it might need to be added
back in at a later date if we decided to use it to style some of the
fields.

Closes #68