Revert "Merge branch 'wsdot-148-wd1'"
This reverts commit bb971a134da423450c4589b1fc8bf759c9a7d4a8, reversing
changes made to 6f0eaf3766c09004e111ca95cf9f12dbb461372f.

With ckeditor 7.x-1.17 the wysiwyg text field within the editor was empty
where there used to be text. Switching to either Filtered HTML or Full HTML
restored the text.

Part of the problem might be that the media integration was removed with this
version.