Update so regional news appears on North and South Central region pages
In the Regional Content pane view's "Contextual filter" for "Regions" needed to
uncheck the "Transform dashes in URL to spaces in term name filter values" option.
Has to do with the term "North Central" being converted to "north-central".
Regional "More news" link not filtering correctly on the main "News" page because of
this.

Work work needs to be done on this.